@@ -68,10 +68,9 @@
     "@types/jest": "^29.2.4",
     "@types/lodash": "^4.14.170",
     "@types/pako": "^1.0.1",
-    "@types/react": "^17.0.8",
+    "@types/react": "^18.0.26",
     "@types/react-color": "^3.0.4",
-    "@types/react-dom": "^17.0.5",
-    "@types/react-redux": "^7.1.9",
+    "@types/react-dom": "^18.0.10",
     "@types/react-router-dom": "^5.1.7",
     "@types/react-select": "^4.0.17",
     "@types/tinycolor2": "^1.4.3",
@@ -108,12 +107,12 @@
     "pako": "^1.0.10",
     "pusher-js": "^7.0.3",
     "qrcode.react": "^3.1.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-color": "^2.19.3",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-password-strength-bar": "^0.3.3",
-    "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.6.1",
     "react-select": "^4.3.1",
     "tinycolor2": "^1.4.2",
     "wegas-ts-api": "link:../../../../../wegas-ts-api/target/wegas-ts-api"
