@@ -24,10 +24,10 @@
         "jsoninput": "github:heigvd/jsoninput",
         "lodash-es": "^4.17.21",
         "prop-types": "^15.7.2",
-        "react": "^17.0.2",
+        "react": "^18.2.0",
         "react-ace": "^6.5.0",
         "react-color": "^2.17.3",
-        "react-dom": "^17.0.2",
+        "react-dom": "^18.2.0",
         "react-select": "^4.3.1",
         "react-sortable-hoc": "^2.0.0",
         "recast": "^0.18.5",
@@ -42,9 +42,9 @@
         "@babel/preset-react": "^7.0.0",
         "@types/lodash-es": "^4.17.4",
         "@types/prop-types": "^15.7.1",
-        "@types/react": "^17.0.19",
+        "@types/react": "^18.0.26",
         "@types/react-color": "^3.0.1",
-        "@types/react-dom": "^17.0.9",
+        "@types/react-dom": "^18.0.10",
         "@types/react-select": "^4.0.17",
         "@types/react-sortable-hoc": "^0.7.1",
         "@types/react-widgets": "^4.3.3",
