@@ -83,13 +83,13 @@
     "proj4": "^2.8.0",
     "pusher-js": "^5.1.1",
     "qrcode.react": "^3.1.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-async-hook": "^4.0.0",
     "react-chartjs-2": "^4.0.1",
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
-    "react-dom": "^17.0.2",
-    "react-reflex": "^3.0.18",
+    "react-dom": "^18.2.0",
+    "react-reflex": "^4.0.9",
     "react-select": "^4.3.1",
     "react-smooth-dnd": "^0.11.1",
     "reactn": "^2.2.6",
@@ -113,8 +113,8 @@
     "@types/lodash-es": "^4.17.4",
     "@types/pako": "^1.0.1",
     "@types/proj4": "^2.5.2",
-    "@types/react": "^17.0.19",
-    "@types/react-dom": "^17.0.9",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.10",
     "@types/react-grid-layout": "^0.17.0",
     "@types/react-select": "^4.0.17",
     "@types/webpack-env": "^1.16.3",
