@@ -13,10 +13,10 @@
     "@babel/polyfill": "^7.4.4",
     "@reduxjs/toolkit": "^1.5.1",
     "chartist": "^0.11.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.6.1",
     "react-select": "^4.3.1",
     "rx": "^4.0.6",
     "wegas-ts-api": "link:../../../../../wegas-ts-api/target/wegas-ts-api"
@@ -26,8 +26,7 @@
     "@babel/preset-env": "^7.5.5",
     "@babel/preset-react": "^7.0.0",
     "@types/chartist": "^0.11.1",
-    "@types/react": "^17.0.19",
-    "@types/react-redux": "^7.1.9",
+    "@types/react": "^18.0.26",
     "babel-eslint": "^10.0.3",
     "babel-loader": "^8.0.6",
     "babel-preset-react": "^6.24.1",
