@@ -1,5 +1,5 @@
 import React from 'react';
-import {HashRouter, Route} from 'react-router-dom';
+import {HashRouter, Route, Routes, useParams} from 'react-router-dom';
 import {Provider} from 'react-redux';
 import {getStore} from './js/init';
 import App from './js/Comps/App';
@@ -7,6 +7,11 @@ import LogIdSelector from './js/Comps/LogIdSelector';
 import Stats from './js/Comps/Stats';
 import ErrorBoundary from './js/Comps/ErrorBoundary';
 
+function StatWrapper() {
+  const { logId: logId } = useParams<{ logId: string }>();
+
+  return <Stats logId={logId || ''}/>
+}
 
 export default function AppRoute(): JSX.Element {
 
@@ -16,8 +21,10 @@ export default function AppRoute(): JSX.Element {
         <HashRouter>
           <div>
             <App />
-            <Route exact path="/" component={LogIdSelector} />
-            <Route component={Stats} path="/:LogId" />
+            <Routes>
+              <Route element={<StatWrapper/>} path="/:logId" />
+              <Route path="/" element={<LogIdSelector/>} />
+            </Routes>
           </div>
         </HashRouter>
       </Provider>
